moves stringifyQuery to helper method
